Organização dos arquivos e adição da função serial_print() em serial.h

+++ src/kernel/serial.c
@@ -63,7 +63,7 @@ void serial_write(uint8_t data) {
  * @param data 
  * @param length 
  */
-void serial_print(uint8_t *data, uint16_t length) {
+void serial_print(uint8_t *data, uint8_t length) {
     for(int i = 0; i < length; i++) {
         serial_write(data[i]);
     }
